requirements.txt
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-core==1.2.28
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.6
python-dotenv==1.2.2
streamlit==1.56.0